Refactor handlers.rs and notification.rs into modular structure
